Moved Word check code to word.ts
