+++ src/app/GameArea.tsx
@@ -13,6 +13,7 @@ import { animate, motion, useMotionValue, useTransform } from "motion/react";
 
 import { KeyInfo, GameData, UIData, ShopEntry, ShopAction } from "./GameData";
 import { GameState, initialGameState } from "./GameState";
+import { nextWordState } from "./word"
 
 import Keyboard, { KeyStatus, KeyMode } from "./Keyboard";
 import ScoreBoard from "./ScoreBoard";
@@ -20,13 +21,6 @@ import DictArea from "./DictArea";
 import InputArea from "./InputArea";
 import Log from "./Log";
 import Shop from "./Shop";
-
-import { Trie } from "./trie/trie";
-/*
-import { TrieNode } from "./trie/trieNode";
-*/
-
-const tdict = Trie.fromArray(GameData.tinydict);
 
 const getKeyMode = (key:string, boughtKeys: Set<string>, repeatableKeys: Set<string>,
    repeatAvailable: boolean, unlockAvailable: boolean, repeatSelectMode: boolean) =>
@@ -55,74 +49,7 @@ const getKeyStatus = (keyInfo: Array<KeyInfo>,
   );
 };
 
-const isPartialWord = (partialWord: string, tdict: Trie) => {
-  const node = tdict.prefixSearch(partialWord);
-  if (node === null)
-    return false;
-
-  return true;
-}
-
-// Is it a word, and am I unable to extend it?
-const isWordTerminal = (word: string, tdict: Trie, maxLength: number): boolean => {
-  if (!tdict.find(word))
-    return false;
-  const node = tdict.prefixSearch(word);
-  if ((word.length == maxLength) ||
-      (node.childrenCount() == 0) )
-    return true;
-
-  return false;
-}
-
-/* 
-  Design : word formation.
-
-  Currently, typed letters can do the following (possibly overlapping) things:
-  -Start a new word
-  -No-op, ie not start a new word, or add to an existing word
-  -Interrupt an incomplete word, and either start a new one, or No-op
-  -Continue ongoing word
-  -Complete a word
-  
-
-  Q : Is state transition completely determined by an op (key, currentWord) -> (currentWord, lastWord) ?
-*/
-
-//TODO : implement maxWordLength properly.
-// Complete next state.  
-const nextWordState = (key: string, currentPartialWord: string, tdict: Trie, maxWordLength: number) => {
-  if (maxWordLength == 0)
-  {
-    return (
-      {
-        currentPartialWord: "",
-        finishedWord: ""
-      });
-  }
-
-  let finishedWord = "";
-  const tentativeWord = currentPartialWord.concat(key);
-
-  if (isPartialWord(tentativeWord, tdict)) {
-    if (isWordTerminal(tentativeWord, tdict, maxWordLength)) {
-      finishedWord = tentativeWord;
-      currentPartialWord = "";
-    }
-    else
-      currentPartialWord = tentativeWord;
-  }
-  else if (tdict.has(currentPartialWord)) {
-    finishedWord = currentPartialWord;
-    currentPartialWord = isPartialWord(key, tdict) ? key : "";
-  }
-
-  return (
-    {
-      currentPartialWord: currentPartialWord,
-      finishedWord: finishedWord
-    });
-}
+/**************************************************************/
 
 const GameArea = () => {
   const [GS, setGS] = useImmer<GameState>(initialGameState);
@@ -198,34 +125,36 @@ const GameArea = () => {
   /**************************************************************************/
   // Keypress handling, scoring, word formation
 
-  const processGlyph = (key: string) => {
-    let buffer = GS.inputBuffer;
-    if (buffer.length == GameData.inputSize) {
-      buffer = buffer.slice(GameData.inputSize / 2, GameData.inputSize);
-    }
-    setGS(gs => {gs.inputBuffer = buffer + key});
-
-    const nextState = nextWordState(key, GS.currentPartialWord, tdict, GS.maxWordSize);
-    setGS(gs => {gs.currentPartialWord = nextState.currentPartialWord});
-    if (nextState.finishedWord != "") {
-      setGS(gs => {gs.lastScoredWord = nextState.finishedWord});
-      setGS(gs => {gs.words++});
-    }
-    setGS(gs => {gs.glyphs++});
-    setGS(gs => {gs.score += GameData.keyScores[key]});
-
-    for (const entry of GameData.shopEntries) {
-      if ((GS.glyphs + 1) >= entry.visibilityPrice)
-        setGS(gs => {gs.visibleShopItems.add(entry.index)});
-    }
-  }
-
   const handleKey = (key: string) => {
     if (GS.boughtKeys.has(key)) {
-      setGS(gs => {gs.lastPressed = key});
-      processGlyph(key);
+      let buffer = GS.inputBuffer;
+      if (buffer.length == GameData.inputSize) {
+        buffer = buffer.slice(GameData.inputSize / 2, GameData.inputSize);
+      }
+
+      const nextState = nextWordState(key, GS.currentPartialWord, GS.maxWordSize);
+      if (nextState.finishedWord) {
+        setGS(gs => {gs.lastScoredWord = nextState.finishedWord;
+                     gs.words++});
+        }
+
+      setGS(gs => {
+        gs.inputBuffer = buffer + key;
+        gs.currentPartialWord = nextState.currentPartialWord;
+        gs.lastPressed = key;
+        gs.glyphs++;
+        gs.score += GameData.keyScores[key]; });
+  
+      // TODO : maybe move this to shop component
+      for (const entry of GameData.shopEntries) {
+        if ((GS.glyphs + 1) >= entry.visibilityPrice)
+          setGS(gs => {gs.visibleShopItems.add(entry.index)});
+      }
     }
   }
+
+  /**************************************************************************/
+  // Keyboard clicks
 
   const keyboardClick = (key: string) => {
     if (GS.repeatSelectMode && GS.repeatableKeys.has(key)) {
